refactor: moved page objects into support folder.
