+++ generators/ionic/templates/cypress/support/pages/entities/_entity.po.ts.ejs
